+++ internal/control_test_engine/ue/nas/message/nas_control/mm_5gs/ul-nas-transport.go
@@ -15,7 +15,7 @@ import (
 
 func UlNasTransport(ue *context.UEContext, requestType uint8) ([]byte, error) {
 
-	pdu := getUlNasTransport_PduSessionEstablishmentRequest(uint8(ue.PduSession.Id), requestType, ue.PduSession.Dnn, &ue.PduSession.Snssai)
+	pdu := getUlNasTransport_PduSessionEstablishmentRequest(ue.PduSession.Id, requestType, ue.PduSession.Dnn, &ue.PduSession.Snssai)
 	if pdu == nil {
 		return nil, fmt.Errorf("Error encoding %s IMSI UE PduSession Establishment Request Msg", ue.UeSecurity.Supi)
 	}
@@ -27,7 +27,7 @@ func UlNasTransport(ue *context.UEContext, requestType uint8) ([]byte, error) {
 	return pdu, nil
 }
 
-func getUlNasTransport_PduSessionEstablishmentRequest(pduSessionId uint8, requestType uint8, dnnString string, sNssai *models.Snssai) (nasPdu []byte) {
+func getUlNasTransport_PduSessionEstablishmentRequest(pduSessionId int64, requestType uint8, dnnString string, sNssai *models.Snssai) (nasPdu []byte) {
 
 	pduSessionEstablishmentRequest := sm_5gs.GetPduSessionEstablishmentRequest(pduSessionId)
